# openapi.yml
openapi: 3.0.3
info:
  title: Space Marine Management API
  version: 1.0.0
  description: API for managing Space Marines, their Coordinates, and Chapters with full CRUD, pagination support, special operations, embedded entities, and export capabilities.
servers:
  - url: http://localhost:8080/api
    description: Local development server
paths:
...
  /special-operations/space-marines/filter-by-weapons:
    get:
      summary: Filter Space Marines by weapon types with pagination
      operationId: filterSpaceMarinesByWeapons
      parameters:
        - name: weaponTypes
          in: query
          required: true
          schema:
            type: array
            items:
              type: string
              enum: [BOLTGUN, HEAVY_BOLTGUN, FLAMER, HEAVY_FLAMER, MULTI_MELTA]
            minItems: 1
            maxItems: 10
          style: form
          explode: true
          description: List of weapon types to filter by (repeat parameter for multiple values)
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
          description: Zero-based page index
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
          description: Number of items per page (max 100)
      responses:
        "200":
          description: Paginated list of filtered Space Marines
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedSpaceMarineResponse"
        "400":
          description: Invalid weapon type or pagination parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /special-operations/chapters/{chapterId}/assign-marine/{marineId}:
    put:
      summary: Assign a Space Marine to a Chapter
      operationId: assignMarineToChapter
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          description: ID of the Chapter to assign the marine to
        - name: marineId
          in: path
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
          description: ID of the Space Marine to assign
      responses:
        "200":
          description: Space Marine successfully assigned to Chapter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpaceMarine"
        "404":
          description: Chapter or Space Marine not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Invalid IDs provided
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ================
  # SPACE MARINES
  # ================
...
components:
  schemas:
...
    PaginatedSpaceMarineResponse:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/SpaceMarine"
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          format: int32
          minimum: 0
...
    SpaceMarine:
      type: object
      required:
        - id
        - name
        - coordinatesId
        - creationDate
        - chapterId
        - health
        - weaponType
      properties:
        id:
          type: integer
          format: int32
          minimum: 1
          description: Unique auto-generated ID (> 0)
        name:
          type: string
          minLength: 1
          description: Non-empty, non-null name
        coordinatesId:
          type: integer
          format: int64
          minimum: 1
          description: ID of associated Coordinates
        creationDate:
          type: string
          format: date-time
          description: Auto-generated creation timestamp
        chapterId:
          type: integer
          format: int64
          minimum: 1
          description: ID of associated Chapter
        health:
          type: integer
          format: int64
          minimum: 1
          description: Health must be > 0
        loyal:
          type: boolean
          nullable: true
        category:
          type: string
          nullable: true
          enum: [AGGRESSOR, INCEPTOR, TACTICAL, CHAPLAIN, APOTHECARY]
          description: Astartes category
        weaponType:
          type: string
          enum: [BOLTGUN, HEAVY_BOLTGUN, FLAMER, HEAVY_FLAMER, MULTI_MELTA]
          description: Type of weapon used by the Space Marine

    # === EMBEDDED SPACE MARINE ===
...
    ErrorResponse:
      type: object
      required:
        - error
        - timestamp
      properties:
        error:
          type: string
          description: Human-readable error message
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the error occurred
      example:
        error: "Invalid weapon types: LASER_CANNON. Valid values: BOLTGUN,HEAVY_BOLTGUN,FLAMER,HEAVY_FLAMER,MULTI_MELTA"
        timestamp: "2025-11-07T14:30:00.000Z"